services/auth/requirements.txt
flask==2.3.2
PyJWT==2.8.0
pyodbc==5.0.0